Reorganize develoop code
